move WikiPages to its own folder
