+++ TASVideos/Pages/Forum/Topics/Index.cshtml.cs
@@ -4,6 +4,7 @@ using Microsoft.EntityFrameworkCore;
 using TASVideos.Core;
 using TASVideos.Core.Services;
 using TASVideos.Core.Services.ExternalMediaPublisher;
+using TASVideos.Core.Services.Wiki;
 using TASVideos.Data;
 using TASVideos.Data.Entity;
 using TASVideos.Data.Entity.Forum;
