+++ TASVideos/Pages/Profile/HomePage.cshtml
@@ -1,11 +1,14 @@
 ﻿@page
-@inject IWikiPages _wikiPages
+@using TASVideos.Core.Services.Wiki
+@using TASVideos.TagHelpers
+@using Microsoft.AspNetCore.Mvc.TagHelpers
+@inject IWikiPages WikiPages
 
 @{
 	ViewData["Title"] = "Home Page";
 	ViewData.AddActivePage(ProfileNavPages.HomePage);
 	var homePage = $"{LinkConstants.HomePages}{User.Name()}";
-	bool pageExists = await _wikiPages.Exists(homePage);
+	bool pageExists = await WikiPages.Exists(homePage);
 }
 <br />
 @if (ViewData.UserHas(PermissionTo.EditHomePage))
