+++ TASVideos/ViewComponents/WikiOrphans.cs
@@ -1,5 +1,5 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using TASVideos.Core.Services;
+using TASVideos.Core.Services.Wiki;
 using TASVideos.WikiEngine;
 
 namespace TASVideos.ViewComponents;
