+++ tests/TASVideos.Core.Tests/Services/LanguagesTests.cs
@@ -1,4 +1,5 @@
 ﻿using TASVideos.Core.Services;
+using TASVideos.Core.Services.Wiki;
 using TASVideos.Data.Entity;
 
 namespace TASVideos.Core.Tests.Services;
