+++ tests/TASVideos.Core.Tests/Services/QueueServiceTests.cs
@@ -1,4 +1,5 @@
 ﻿using TASVideos.Core.Services;
+using TASVideos.Core.Services.Wiki;
 using TASVideos.Core.Services.Youtube;
 using TASVideos.Core.Settings;
 using TASVideos.Data.Entity;
