+++ tests/TASVideos.Core.Tests/Services/UserFilesTests.cs
@@ -1,4 +1,5 @@
 ﻿using TASVideos.Core.Services;
+using TASVideos.Core.Services.Wiki;
 using TASVideos.Data.Entity;
 using TASVideos.MovieParsers;
 using TASVideos.MovieParsers.Result;
