+++ tests/TASVideos.Core.Tests/Services/WikiPagesTests.cs
@@ -1,4 +1,4 @@
-﻿using TASVideos.Core.Services;
+﻿using TASVideos.Core.Services.Wiki;
 using TASVideos.Data.Entity;
 
 namespace TASVideos.Core.Tests.Services;
